classes

+++ engine/src/engine/range/Range.java
@@ -1,30 +1,30 @@
 package engine.range;
 
 public class Range {
-    private int to = 0;
-    private int from = 0;
+    private float to = 0;
+    private float from = 0;
 
     public Range() {
     }
 
-    public Range(int to, int from) {
+    public Range(float to, float from) {
         this.to = to;
         this.from = from;
     }
 
-    public void setTo(int to) {
+    public void setTo(float to) {
         this.to = to;
     }
 
-    public void setFrom(int from) {
+    public void setFrom(float from) {
         this.from = from;
     }
 
-    public int getFrom() {
+    public float getFrom() {
         return from;
     }
 
-    public int getTo() {
+    public float getTo() {
         return to;
     }
 }
